code/m4estro/LambdaPY/graph_manager/builder_service/local/requirements.txt
notebook
ipykernel
numpy
pandas
matplotlib
python-igraph
pycairo